Implementation for path matching.
Updated the GetEndpointsForPath to return an array of Endpoints rather
than a single endpoint. The first endpoint might not have the correct
methods on it, so may need to process further ones.

Updated all path splitting to return an array with a single empty
string at a minimum. This is the base path to be matched. At every
level of the response heirarchy, there will always be at least an
empty string for the path.

Added tests for the GetEndpointsForPath method.