Can't deseralise interfaces (obviously), remove them
Not sure what I was thinking here, that was never going to work. I've
removed the interfaces for now and made the methods that I want to mock
virtual. Not sure how I feel about this really, I'm a bit uneasy about
it at the moment but I'll do a bit of research and see if it's okay.
It's working for now though.